+++ NEWS.md
@@ -29,6 +29,8 @@
 
 * Fixed a bug in `tableby()` relating to a weird edge case when a by-level is "Total" and the total label is set to something else.
 
+* Added a feature to `tableby.control()` to allow for the total column to be moved before the other columns. (#320)
+
 # arsenal v3.5.0
 
 * Change R requirement to >= 3.4.0.